refactor(root): extract tests and improve header parsing clarity
- Extract all tests from block.rs into separate block_tests.rs module for
  better organization and maintainability
- Refactor RootVersion match statement to use dedicated header parsing
  functions: parse_v1_header, parse_v2v3_header, parse_v4_header
- Rename V2BlockHeader fields to clarify structure:
  - content_flags → content_flags_1
  - unk2 → content_flags_2
  - unk3 → content_flags_3
  - Add reconstruct_v2_content_flags helper to document flag reassembly
- Improve documentation: clarify V2 header introduced in build 11.1.0.58221
  and explain how 3 split fields reconstruct the full content_flags value
- Fix ContentFlags::NO_NAME_HASH position: bit 13 → bit 28 (0x1000_0000)
